+++ lib/vagrant-vbguest/config.rb
@@ -3,7 +3,7 @@ module VagrantVbguest
   class Config <( Vagrant::VERSION < "1.1.0" ? Vagrant::Config::Base : Vagrant.plugin("2", :config) )
 
     module Attributes
-      attr_accessor :iso_path, :auto_update, :auto_reboot, :no_install, :no_remote, :installer
+      attr_accessor :iso_path, :auto_update, :auto_reboot, :no_install, :no_remote, :installer, :installer_arguments
     end
 
     class << self
@@ -13,6 +13,7 @@ module VagrantVbguest
       def auto_reboot; @auto_reboot.nil? ? true  : @auto_reboot end
       def no_install;  @no_install.nil?  ? false : @no_install  end
       def no_remote;   @no_remote.nil?   ? false : @no_remote   end
+      def installer_arguments; @installer_arguments.nil? ? '--nox11' : @installer_arguments end
 
       def iso_path
         return nil if !@iso_path || @iso_path == :auto
@@ -26,6 +27,7 @@ module VagrantVbguest
     def auto_reboot; @auto_reboot.nil? ? self.class.auto_reboot : @auto_reboot end
     def no_install;  @no_install.nil?  ? self.class.no_install  : @no_install  end
     def no_remote;   @no_remote.nil?   ? self.class.no_remote   : @no_remote   end
+    def installer_arguments; @installer_arguments.nil? ? self.class.installer_arguments : @installer_arguments end
 
     def iso_path
       return self.class.iso_path if !@iso_path || @iso_path == :auto
@@ -36,6 +38,7 @@ module VagrantVbguest
     def to_hash
       {
         :installer => installer,
+        :installer_arguments => installer_arguments,
         :iso_path => iso_path,
         :auto_update => auto_update,
         :auto_reboot => auto_reboot,